Refactor code, and fix a mistake
